Split StdFaceMap controls to own file
